+++ CHANGES.rst
@@ -62,6 +62,7 @@ Changelog
 - Add IDE feature to generate view from form with selected fields [quang]
 - Fix to export function, causing additional lines in exported python [djay]
 - Add search form functionality [quang]
+- Bug fix for extra space on triple quote statement [ivanteoh]
 
 2.0b4 (2016-05-27)
 ------------------